Add a per-package mode for older protoc plugins
A few years back, none of the Go plugins could handle multiple package
values in the same compliation. It wasn't until the rewrite of
golang/protobuf that multi-package builds started working.

Unfortunately, there are still some plugins that never upgraded. This
patch adds a flag that batches targets by package and compiles each
batch separately. When the flag is off then it should work the same as
it did before.