More mgdl-gen features (#42)
* Add support for enum parameters in SDK 
* Add native API interface generation
* Add lazy package and import declarations in generatedFile
* Auto import for any reference to a type from another module
* Topological sorting of protoc plugin inputs

API generation is supported by adding another parameter and re-arranging the generator code as a struct so it can reference the parameter state. Nearly all the generator code is identical but may not show that way on the diff because I also moved it at the same time.

The import support in mgdl-gen is a little hack-ish. For now, I just pass the generatedFile and the current module UID to the genType method so that it can have the side-effect of adding imports. It also relies on the proto annotation for Go package.

The topo sort came up when I had something import that, I guess for the first time, wasn't incidentally topo sorted due to the directory layout. The CodeGeneratorRequest specifies that the input will be pre-sorted and it appears that protoc-gen-go relies on this behavior. I can see how this provides for some optimizations in plugins but I want to think through it more before making it part of the compiler contract in general. For now, I copy/pasted some topo sort code I have for another project into the proto descriptor conversion code so that protoc plugins get what they expect.