Fix more proto map type resolution issues (#45)
I found a new variation on the map type resolution issue. When compiling
a proto file including a map definition I ended up with a type reference
of <StructName>.<MapEntryName> instead of <MapEntryName>. This resulted
in a miss on the addMapEntryQualifications because it was looking for
unqualified references to the entry type. Now, the search includes all
variations of the entry type from a rooted, qualified name to the
unqualified base name. Additionally, the code is updated to account for
all those variations when building back up the fully qualified path.